--- components/inspectit-ocelot-configurationserver-ui/pages/abc.js
@@ -1,10 +1,0 @@
-import Link from '../components/basics/Link'
-
-export default () => {
-    return (
-        <div>
-            hello
-            <Link href="/"><a>index</a></Link>
-        </div>
-    )
-}
