+++ components/inspectit-ocelot-configurationserver-ui/redux/ducks/clock/types.js
@@ -1,3 +1,4 @@
+export const INIT = 'clock/INIT';
 export const TICK = 'clock/TICK';
 export const INCREMENT = 'clock/INCREMENT';
 export const DECREMENT = 'clock/DECREMENT';
